boot/boottest: helpers to set boot kernel and base
This makes it so that only tests that care about the boot mechanics
need to care about the details of it.

These tests might be better off be split and part of them moved into
boot, as there's little reason for them to be where they are caring
for the level they have (e.g. devicemanager shouldn't need to care
about the vagaries of boot rollback, probably).

But that's work for another PR.