tests: measure testbed for leaking mountinfo entries
Tests can routinely leak system wide side effects, like mount tables.
This patch adds the `testbed-tool` executable that with two
sub-commands, one for setting baseline state and another for comparing
the current state to the baseline state.

The tool is applied in a very specific spot in the prepare/restore
logic, because that is where it seems to be mostly effective. The
location is chosen because prepare/restore is really mixed up in snapd
test suite and naive approach of setting the baseline after prepare and
comparing it to the current state after restore does not really
function.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>