tests: remove core18 if installed by tests
When tests run on ubuntu core, the automatic cleanup performed is
different and imperfect. In particular it is very easy to "leak"
additional base snaps, like core18.

Before we have a better mechanism available, remove core18 while
restoring a test case that did not have core18 before the test executed.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>