Improve how the system is restored when the upgrade-from-2.15 test fails
This fix is to make sure when the upgrade-from-2.15 test fails the
systems restores snapd and its dependencies properly.

To reproduce the error it is needed to add an "exit 1" at the begining
of the execution part, and the after this test failure all the following
tests will fail as well.

This fix was tested making the test fail before the old snapd is
installed and after that as well.