tests: be less verbose when checking service status
Our classic system test reset code is very chatty, as it will attempt,
up to 20 times, to check the status of snapd.service. In each time it
really only cares about a particular property but it doesn't hide the
output from systemctl status. When debugging other issues I found that I
was going through pages of this output so here's a small way to make the
test log shorter.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>