Sync journal log before start the test
Sporadically some logs were being logged late (as part of the next
test). As the catalog-update test is checking if the refresh is verbose
or not, some messages could affect that check as in the following error:

+ echo 'Catalog update is doing verbose http logging (it should not).'
Catalog update is doing verbose http logging (it should not).
+ exit 1

The idea is to sync and flush the logs before start the new test as we
are doing before query the logs.

See full test error:
https://paste.ubuntu.com/p/DvpmHKcHtH/