many: introduce confinement options failsafe flag
This patch introduces a new internal confinement option flag called failsafe
mode. When a security setup fails it is attempted again in failsafe mode. This
can be used to mitigate impact of both known and unknown issues in the
interface system that, otherwise, might result in total system failure.

The first consumer of the failsafe mode is the mount backend. When the flag is
asserted the mount namespace will be discarded ahead of the setup. This allows
to nullify effects of corrupted state of a mount namespace preventing changes
from being applied.

This is both a workaround and a reliability idea. I'm sharing it with Ondra for
some experiments.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>