overlord/snapstate: use reflect to improve snapstate.Get
Based on [some profiling I did ages ago][1], loading a single element
from a json object using reflect to create an ad-hoc parser is more
efficient in both memory and speed than our current approach of using
a map[string]*json.RawMessage.

This implements this style of loader for one of the places where it
would help, in snapstate.

[1]: https://gist.github.com/chipaca/10ecdf364db962c9b5ee590f0f610b2a