tests: add netplan-apply spread test
This spread test installs netplan outside of confinement (for the D-Bus service)
as well as inside a strictly confined snap which plugs network-setup-control to
test running netplan apply inside the snap (but really resolving to running that
command outside of confinement).

Note that this test currently manually installs the netplan D-Bus service on the
spread system until the SRU is finished to allow just installing the debian
package and getting the D-Bus service by default.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>