tests: measure behavior of the device cgroup
This test checks the interaction of systemd and snap-confine with
regards to the device cgroup. Since the 2.0 release, snap-confine is
using the device cgroup to constrain access to character and block
devices. Effectively, snap-confine is stealing the process away from
systemd by moving it from the cgroup managed by systemd to one managed
by snapd.

Note that this model can only work in cgroup v1, where a process can at
a given time inhabit multiple separate cgroups.

Not all processes us the snapd-made device cgroup. If a snap is placed
in devmode or uses no interfaces that would toggle udev to tag a device
to add to the cgroup, then no additional cgroup is created or
configured.

This model lead to an overlooked conflict between how systemd and snapd
configure a process that started out without udev tagged devices, and
thus without a snapd-managed cgroup, is connected to an interface that
relies on tagging for precise confinement.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>