usersession/agent: use background context when stopping the agent
When agent http.Serve() exits, the tomb will cancel all child contexts. Using a
child context of a tomb  (as obtained from s.tomb.Context(nil)), exposes us to a
race between http.Shutdown() observing all connections as closed, and context
being canceled by the tomb.

It is ok to use the background context, because Shutdown() is a method of the
http.Server being closed, thus it will correctly observe the server becoming
quiescent and exit or the context will hit the timeout and a DeadlineExceeded
will be obtained.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>