tests: make interfaces-contacts-services cleanup more robust
We saw a failure in the test cleanup of interfaces-contacts-service.
The error is:
```
+ rm -rf /home/gopath/src/github.com/snapcore/snapd/tests/main/interfaces-contacts-service/xdg
rm: cannot remove '/home/gopath/src/github.com/snapcore/snapd/tests/main/interfaces-contacts-service/xdg/share/gvfs-metadata': Directory not empty
```
We already have code that deals with this but apparently its not
sufficient (anymore?).

This PR adds some debug output to ensure we have more data if
this happens again. But it also stops the gvfs-metadata user
service if its available to ensure we don't get into funny races
with systemd that may try to start/restart the service.