sanity: unmount squashfs with --lazy
While performing the mount namespace cleanup tests I noticed a
failure resulting from leaked mount point:

/dev/loop1 /tmp/sanity-mountpoint-773821774 squashfs rw,relatime ro

This looks like the sanity check for squashfs support has failed. While
I have no idea why it has failed and was unable to reproduce it easily,
it may be because something went to inspect the new mount point at the
exact moment we wanted to unmount it.

To be more robust in this case, use umount --lazy, so that we don't fail
in case there are open file descriptors.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>