tests: show why sbuild failed
The sbuild test seems to fail sometimes, but not due to timeout that
kills the test. When that happens test logs contain absolutely nothing
useful.

Maciej suggested that we can look at the tail of the sbuild logs to do
so. In a practical run invoked with --shell-after there are two log
files:

    snapd_2.40-2_amd64-2019-08-13T09:46:14Z.build
    snapd_2.40-2_amd64.build -> snapd_2.40-2_amd64-2019-08-13T09:46:14Z.build

I don't want to be in the business of extracting that version so I just
take the one with the Z.build suffix.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>