cmd/snap-update-ns: check for trespassing using clean path
When the trespassing detector described in the previous patch is used to
check if a directory is safe to write to, it iterates over past mount
namespace changes to detect presence of a tmpfs mounted by
snap-update-ns. The check is based on path equality since the check is
performed for each directory depth, it will correctly match a tmpfs
mounted at the parent directory.

The check is invoked from MkPrefix, which is invoked from MkdirAll and
MkfileAll. The purpose of the prefix function is to create everything up
to the point where then leaf (either a directory or a file) can be
created. The prefix function is provided with the path to create.
Internally it uses a path iterator to go over the components of the path
one by one. For each of the components it delegates the construction to
MkDir, giving it the directory name to create, the parent directory
represented as a file descriptor as well as the parent directory
represented as a string. The string is only used by the trespassing
detector.

The error manifests itself when the parent directory represented as a
string is not clean, because it will not match the strict requirements
of the trespassing detector. This causes incorrect trespassing
violations that bubble up and cause failures in mount namespace changes.
This patch fixes it by providing a clean path to the trespassing
detector.

Fixes: https://bugs.launchpad.net/snapd/+bug/1808821
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>