client: do not leak retry timers
Each request to snapd gets a timer that marks the upper limit on the complete,
potentially retried, request loop.

However, the timer is set up using time.After(), which creates timer that only
gets garbage collected once it expires. With the aggressive poll times in the
cli (100ms), it is possible there is quite some number of the timers in the
client while waiting for snap operations to complete.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>