tests: rewrite "retry" command as retry-tool
Having pushed Sergio towards writing the command in Python instead of
shell I think it's on me to make sure it lands and looks right. This
patch rewrites the utility, with extra usability touches and, mainly,
correct semantics.

I added a quick spread test to showcase the features as well.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>