[WIP] interfaces/serial-port: support pci bus serial-port with HotplugKey()
Support defining a HotplugKey that uses DEVPATH as part of it's udev attribute
seeds in the key to uniquely identify the device. If the bus of the serial-port
isn't pci, or we still don't find a sufficient pair of vendor / model attributes
then bail out and try to use the default key instead.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>