tests: fix user-libnss test
For unknown (and probably silly) reasons useradd segfaults if
the "gshadow" file is missing. It seems to ignore it when it
is missing when it is opening files but then tries to fclose()
it later anyway.

The workaround is to create an empty /var/lib/extrausers/gshadow
file (and shadow for good measure).