cmd/snap: fix remote snap info for parallel installed snaps
When running `snap info` for a parallel installed snap, the code would ask the
store for information using the actual instance name (<snap>_<key>) instead of
just the snap name. This lead for remote info, such as channels, not being
displayed in the output.

Use the snap name when looking up snaps in the store.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>