tests: clean up after NFS tests
The test was leaking two mount points: /var/lib/nfs/rpc_pipefs and
/proc/fs/nfsd. In addition because of how the test operates, files
placed in the current working directory can get "lost" somehow. As a
precaution I moved all of the helper files to /tmp.

Note that the entire test is disabled on Fedora, CentOS, openSUSE and
Ubuntu Core so there is still some "dead code" in terms of switch
statements that are not executed. I tried to do my best to keep the dead
code as useful, to the future developers, as I could.

There are small unrelated changes as well, like how /home is unmounted
or a FIXME note about test clobbering a file on Fedora.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>