sanity: report proper errror when fuse is needed but not available
During the spread run for PR#7286 we noticed that for some systems
our sanity check in lxd is not helpful. When a container does
not have fuse at all we try to mount the fusesnap filesystem
but because there is no fuse the error message is generic and
confusing. To fix this this PR adds a check if fuse is needed
and if the fuse mount binary is missing.