tests/overlord/snapstate: update lists of fake ops expected
This takes into account the new fake ops, restore-disabled-services and
current-snap-service-states, which should be called after a successful link-snap
op, and before an unlink-snap op, respectively unless we are undoing an
operation and there are no previous revisions to go back to. If there are no
previous revisions to go back to, then we shouldn't call these since there's no
state that we could hypothetically lose by not tracking what services were
disabled.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>