tests/overlord/snapstate: add tests for RestoreDisabledServices rename logic
This ensures that if disabled services are renamed (i.e. the same service name
no longer appears) in a future revision, then we keep that service name in the
list of disabled services in the state so that if we ever go back to a revision
that has those services, we will re-disable them.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>