mkversion.sh: fix version from git checkouts
The mkversion script is used during the build of the snapd snap.
Today its a bit naive and just uses "git describe" there to
generate the version number. The downside of this is that for
pre-releases like 2.41~pre1 we may end up with a version number
lower than the version in debian/changelog if the pre release
does not have a git tag. We historically don't tag pre/rc
releases and doing so would be a bit messy (IMO) so we need
a less naive approach to generate the version number.

This PR changes the mkversion.sh script to use something
similar to what a packaging recipe does to construct a version.

On the current master this will yield:
 2.41~pre1+git906.g45df4b1
instead of
 2.40+git872.gf9b531f

Which is what we want.