tests: use user-tool to remove test user in the non-home test
Sometimes the tests for the non-home user fail with:
```
+ deluser --remove-home jim
Looking for files to backup/remove ...
Removing files ...
Removing user `jim' ...
Warning: group `jim' has no more members.
userdel: user jim is currently used by process 24139
/usr/sbin/deluser: `/sbin/userdel jim' returned error code 8. Exiting.
```

we don't really care if there is a process left so this PR switches
to use the new "user-tool remove-with-group" which will ignore
if the user is still in use.