overlord/snapstate: add a defer to call UnlinkSnap on error
This ensures we don't forget to cleanup at any point later in this method, as
the undo handler for undoLinkSnap only runs if doLinkSnap successfully finishes.

Also don't save the disabled services if we fail to run LinkSnap, as there is no
loss of disabled service states between a successful unlink and an unsuccessful
link.

Finally, update unit tests with this change.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>