tests: fix gadget-update-pc test leaking snaps
This test was long on the list of well-known tests that leak stuff. In
particular, this test is leaking revisions of the pc gadget, that remain
mounted even after their snaps are deleted from disk.

The fix, in hindsight, is rather simple and should have been obvious
earlier. The entire problem is that attempting to re-install the
original snap and get the perfect outcome is hard (for example, loopback
devices may change). Instead of doing that, just _retain_ the original
snap, keep the other snaps while we are testing, and revert to the
original revision afterwards.

This passes as clean on the long-lived, in progress mount leak detector
branch.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>