snapstate: validate all system-usernames before creating them
The current checkSystemUsernames code is doing check and creation
at the same time. This may lead to side-effects like creating
users even if the snap cannot be installed because it contains
unsupported/illegal system-users. This showed up in the tests
when sometimes the "system-usernames-illegal" test would create
a "snap_daemon" user even though the snap itself could not be
installed.

This PR fixes it by splitting the check and creation into
two separate steps. The creation only happens if all system-users
are valid.