snapstate,store: check assumes early before downloading the snap
This PR ensures we check for the "assumes" of a snap eary. Instead
of downloading it first and then checking we will now check once
we get the snap info data from the store and bail early.

We already have all the needed code, we just need to populate
the data we get from the store correctly. Plus this adds a spread
tests to ensure this actually works correctly.