cmd:libsnap: add sc_panic, a layer below die()
When die() dies, it doesn't give us much choice as to how to go.  While
reviewing a pull request https://github.com/snapcore/snapd/pull/7358 we
realized that system-shutdown was defining die twice, once from
libsnap-confine-private, using the regular die-and-exit code, and once
internally with custom die-with-reboot.

This isn't great. We should just have one function and allow a process
to say how it wants to customize death.

This is what sc_panic() is. It's a drop-in replacement to die() (in
fact, die will be a thin wrapper around it). On top of die it offers a
pair of customization points: one for how to exit (e.g. you can call
reboot) and one for how to say something failed (you can print to
whatever is appropriate, e.g. to a file).

This patch contains just the new panic function along with documentation
and tests.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>