overlord/snapstate: add migration function to fix invalid channel spec
Previously, snapd performed no validation of the channel spec on refresh
and install and just passed the string to the store for validation,
however now we have to handle model constraints amongst other things and
so we check the channel spec on refresh/update. This is problematic for
clients that may have installed snaps using an unsupported channel spec
with a leading "/" as in "/<risk>" which either implied "latest/<risk>"
or as in "/<track>" which implied just "<track>".

This patch fixes the problem "on the fly" when we encounter such a
channel spec in case it exists in some deployments and corrects it by
removing the leading "/" in order to keep the same expected behavior
such that "/stable" gets changed to "stable" which has a defined
semantic around it.

Notably, we do this before calling resolveChannel, because we don't want
to have that function grow too complicated so it can be more generally
useful and instead just patch the snap state before that function gets
called so it "just works" by the time we get there.

The migration function also returns the new channel so that the in 
memory snapst can be updated to match what's now in the state so there
are no inconsistencies later on, otherwise we will get change conflicts.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>