snap-seccomp: do not rename github.com/mvo5/libseccomp-golang
By avoiding this rename the patches in the debian packaging
will still cleanly apply. The rename is not strictly needed,
it adds a bit of readbility as its not clear that libseccomp-golang
has the "seccomp" package name but I think its ok to leave things
as they are especially in a refactor like the current one where
its about renaming ubuntu*architecture dpkg*archicture.