snapstate: add comment to checkVersion vs strutil.VersionCompare
We are using a custom version compare in snapstate.checkVersion()
which is subtly different from the way strutil.VersionCompare
works (which we use in the other places where we compare versions).

The difference is that checkVersion() does not look at "~". So
an "assumes: [snapd2.41]" is satisfied by snapd version
2.41~pre1.

We need to decide if that is what we want, right now this PR just
adds a test to make this explicit and adds a note about it.

If we want to change it we can use
https://github.com/snapcore/snapd/pull/7344