osutil: make flock test more robust
I noticed that flock unit test has failed in a random run.

    -----
    FAIL: flock_test.go:122: flockSuite.TestLockUnlockNonblockingWorks

    flock_test.go:144:
        c.Assert(lock.TryLock(), Equals, osutil.ErrAlreadyLocked)
    ... obtained = nil
    ... expected *errors.errorString = &errors.errorString{s:"cannot acquire lock, already locked"} ("cannot acquire lock, already locked")
    ... runtime error: invalid memory address or nil pointer dereference

    OOPS: 236 passed, 1 FAILED
    --- FAIL: Test (1.02s)
    FAIL
    FAIL	github.com/snapcore/snapd/osutil	1.023s
    -----

Quick inspection doesn't reveal the cause but there were some red flags.
The synchronization between the test launching flock in another process
and the test itself is racy. It had up to 10ms to succeed. This may be
too short so I adjusted that to give it ample time to run on a busier
system.

What is more mysterious is the error itself, traversal of a nil pointer.
This seems to suggest that we return nil, nil - that is fail without
reporting an error - something NewFileLock just cannot do.

I've added an extra assertion for that so that if it happens, it is more
easy to deduce the cause.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>