cmd/snap: don't append / to snap name just because a dir exists
Before this change, when a user did `snap info foo` and a directory
existed, a `/` was added to `foo` as part of path normalization.

That should only be done to the function that sets up the 'disk' snap,
not for every use of the snap name.

This change fixes that.