devicestate: add support for base->base remodel
This PR adds support for remodel between different base snaps
like core18 to core20. Bases have to be compatible in some
way, i.e. be able to read the old /writable content. The
code does not enforce this in itself.

It does not address "core" -> "$base" remodels yet but its an
important step in that direction and to support UC16->UC18
remodels.