cmd/snap-confine: unmount /writable from snap view
On core systems we establish an elaborate mount namespace, comprised
of the bootable base snap and a lot of bind mounts to the writable
partition.

The writable partition, in its full glory is visible both /writable
and /var/lib/snapd/hostfs/writable. Each view contains /writable/snap/...
coupled with all the snaps visible there.

None of the snap interfaces grant access to actual /writable itself. In
addition the data visible there is totally redundant, because the snaps
are already visible in /snap anyway.

The snaps are visible there because of propagation established by systemd
in the initial mount namespace.

This patch has two changes:
 - It adjust snap-confine to detach /writable/
   and /var/lib/snapd/hostfs/writable/ and anything underneath
   from the mount namespace used by snaps. There's a small permission
   change to make this possible.
 - It adjusts the test data in the mount-ns test to reflect this.

Overall, coupled with other changes to this area, this patch reduces the
complexity of the mount namespace used on ubuntu-core.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>