overlord/snapstate/policy, etc: introduce policy, move canRemove to it
This is the first step towards having a fully fledged snapstate/policy
package where logic around type and model decisions can be placed.

For now it's connected in a rather hackish way (which we're already
using elsewhere) although the expectation is that this could change
with some major refactoring of snapstate itself.

This first bit of work moves canRemove.

The tests are a little messed up, because they are straight
translations of the existing ones which were similarly confused. I'd
like to reorganise them so they make more sense though.