packaging: remove obsolete usr.lib.snapd.snap-confine in postinst
We renamed the apparmor profile for snap-confine to
usr.lib.snapd.snap-confine.real a while ago. Unfortunately due to
a buggy upgrade some people still have the old file and that can
cause real issues like:
https://forum.snapcraft.io/t/telegram-snap-fails-to-start/13132
when this file is used instead of the correct one.

This PR uses a rather aggressive approach to get rid of the file.
Fwiw, I wish we could ship these files in something else than /etc
they are not really conffiles anyway IMO, ppl fiddling with them
have an extremly high chance of breaking snapd this way.