gadget: accept system-seed role and ubuntu-data label
Core 20 introduces the recovery partition and its system-seed role,
and changes the writable system-data partition label to ubuntu-data.
Add minimal support to gadget to recognize the new role and label
to allow a volume definition to be correctly parsed.

We're currently allowing both writable and ubuntu-data labels to be
used for the system-data partition. In a future change we should
detect what is the appropriate label based on the system version.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>