tests/main/snap-run: disable strace test cases on Arch
We have observed issues running strace on Arch. This is likely related to a
known problem with strace and go program getting into a deadlock on the kernel
side.

The proposed patch https://lore.kernel.org/patchwork/patch/719314/ was not
applied in the mainline. Looks like there is no clear fix for the problem yet.

On our side, the problem can be observed when snap-exec gets stuck. The kernel
task dump https://paste.ubuntu.com/p/8nVzj8Sqfq/ reveals that boths strace and
snap-exec are in D state (TASK_UNINTERRUPTIBLE?).

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>