overlord/snapstate: add SetTaskSnapSetup + unit tests
Add a new function SetTaskSnapSetup which will update the SnapSetup stored in a
task if that task is the first task (and hence is the snap-setup-task for all
future tasks in that change), or it will update the SnapSetup stored in that
task's snap-setup-task (also hence updating it for all future tasks in that
change).

Also add unit tests for this function to a new handlers_test.go test file, since
this function is generic for any handler and not for the specific ones in
handlers_*_test.go currently.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>