cmd: add `snap debug bootvars` that dumps the current bootvars
The new debug helper `snap debug bootvars` will output the
current bootloader variables. This is useful on core18 systems
that do not have the grub-editenv/fw_getenv commands and for
bootloaders like little-kernel that do not have any tools for
this.

This will primarely be used in tests like:
https://github.com/snapcore/snapd/pull/7508

Or the test in
https://github.com/snapcore/snapd/pull/7438