daemon: return "snapname_rev.snap" style when using /v2/download
This PR changs the /v2/download endpoint to return a more useful
filename for the downloaded snap. It uses the form
"snapName_rev.snap", e.g. "core_1234.snap". The old code was
just using the snap name which is a bit too limiting.

This will help with the `snap download` via the snapd API code.