wrappers: update the code for generating timer schedules
With the changes in the interpretation of schedules, the code needs to be
updated and new edge cases need to be handled. Extend the test cases to cover
schedules that are currently accepted.

Note that due to inability to express the schedule as systemd.time(8) format, we
still refrain to having systemd fire the timer more frequently than required,
with gating implemented on the `snap run --timer=...` side.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>