gadget: also recognize ubuntu-data label for writable
In the London sprint it was decided that the writable partition will be
labeled ubuntu-data in core20. A system-wide way to detect if snapd is
running on a core20 or legacy system is yet to be determined, so in order
to allow cmd/snap-install to run correctly, temporarily allow both labels.

This-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>