cmd/libsnap: use cgroup.procs instead of tasks
To associate a thread with a cgroup v1 hierarchy one needs to write the
task id of that thread to the "tasks" file. Since snap-confine is
single-threaded this has the same semantics as writing the pid to the
"cgroup.tasks" file.

The advantage of cgroup.procs is that is common across both v1 and v2
so let's use that instead.

This change is applied to: freezer, pids and devices cgroups.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>