client: add KnownOptions to Know() and support remote assertions
The daemon can query remote assertions from the store now. This
PR adds support for this to the client package via a new
KnownOptions parameter to the client.Known() call. This breaks
API which should be ok given that its a internal client API.

If this is a problem we can use #7573