recovery-tool: add sfdisk wrapper
Implement sfdisk support conforming to a minimal protocol established
by the partitionTool interface in such a way that, given a gadget volume
description and a list of existing structures, partitions for the
missing structures are created.

Information on the device node for each created structure is returned
so the caller can do further work on the partitions such as create
filesystems or encrypted devices.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>