tests: when using lxd, remove it in restore
Workaround a problem with snap service stop commands and make sure to remove the
lxd snap in restore.

In the case of lxd snap, the snap specifies a stop command, which in turn is via
the relevant entry in the systemd service unit:

```
ExecStop=/usr/bin/snap run --command=stop lxd.daemon
```

However, the cleanup when done via a postrm deb hook, excutes when /usr/bin/snap
no longer exists. In this case, the lxd packaging is unable to perform a cleanup
and mount entries cluttering the host namespace are left behind.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>