Configure the journald service for core systems
This change makes the same journal configuration for classic and core
systems.

Previously the configuration for the journald service was being done on
prepare_project function (prepare-restore.sh), but when the target
system is core, the prepare_project function is executed on
ubuntu-16.04-64 for core16 or ubuntu-18.04-64 for core18. It happens
because the target ubuntu-core system is created during the first test
suite preparation which happens after the prepare_project function
execution.

The idea is to configure the journald service after the ubuntu-core
system is booted.