cmd/snap-confine: remove leftover condition from capability world
At some point in time we attempted to use capabilities and a non-root
snap-confine but this work was abandoned. Recently I removed most of
the remains of that but didn't notice this one: the code that prepares
classic and non-classic execution environment was conditional on
geteuid() == 0, even though a few lines above we check and die if that
same condition is not met.

As such, remove the condition and adjust indent the affected code.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>